Redo build system to use Automake and closer to a current layout
Move configure.in to configure.ac and rewrite for Autoconf 2.61 with
proper quoting.  Add the m4 fragments from remctl 2.13 and remove the
old aclocal.m4 file.  Rewrite the build system using Automake and
remove Makefile.in.  Fix the leading comments in the m4 fragments.  Add
a bunch of Autoconf tests we'll be using soon.

Rename replace to portable and add portable/dummy.c to avoid empty
libraries.  Rename the test suite accordingly.

Rename the test directory from t to tests to match other projects and to
be clearer.

Fix the prototype of xwritev in include/librutil.h to not use an array
of undeclared structs.

Stop running the runtests tests here, as that's going to move to the
c-tap-driver distribution.