Separate out the message functions that can call exit
It's useful for automated analysis to scan for calls to exit in shared
libraries, since this is normally an error.  However, combining the
message handlers that don't call exit with those that do in the same
file causes the shared libraries to contain code that calls exit, even
if it's never called by the library.

Separate the die and sysdie functions and supporting code into a
separate file in the hope that the linker will then be smart enough to
omit that object code from the shared library.