Change the error message stripping in ok_program
Rather than stripping everything after the second colon, strip everything
after the first colon.  Hopefully this will work everywhere; if not, I'll
have to provide two functions.  This is what I needed for remctl.

This change is only to synchronize with C TAP Harness.  The feature isn't
used in the test suite of this package.