krb5-config cleanup for krb5.m4 and gssapi.m4
If krb5-config produces results that don't work for Kerberos v5 of
GSS-API probes, fall back on manual library probing rather than just
failing.  This fixes problems with GSS-API probes on Solaris where
krb5-config is present but doesn't know about GSS-API.  Thanks to Andy
Cobaugh for the report.

Suppress error output from krb5-config in Kerberos v5 and GSS-API
probes so that it's not mixed in with configure output.

Do not try to use a non-executable krb5-config for GSS-API probes and
prefer the KRB5_CONFIG environment variable over a path constructed
from the --with-gssapi argument.