Revert split of message functions that call exit
The previous split of message functions that could call exit to a
separate source file was indended so that this source file would
not be linked with shared libraries and produce false positives with
code analysis tools.  Unfortunately, this didn't work; libtool and
the linker still include the functions.  Recombine the files.

This reverts commit c41fe98b4b56b67096302f60ae03a6a6fda25f7e.
This reverts commit 02f81a0889513a2ef4589c185ec1d266d2737f97.