Add a libkafs replacement
Add a replacement for the libkafs library for AFS PAG and token
manipulation.  This is a minimal version of the library with no
external dependencies on Linux and only minimal AFS dependencies on
other platforms that implements just enough of libkafs functionality
to support pam-afs-session and kstart.  This is a somewhat-modified
version of the kafs replacement files in kstart 3.14 that uses a
separate library directory and Automake conditionals rather than
AC_LIBOBJ and the portable directory and supports more specific
configure options.

This builds but is as yet untested.  A test suite will be coming in
subsequent commits.