Redo the kafs build system
Rather than using Automake conditionals to control compilation of the
separate AFS system call interfaces, include the *.c files into the
main implementation file using preprocessor defines to control which
is used.  This produces simpler and cleaner Makefiles.

Rename api.c to kafs.c now that it's the only public interface to the
library and most of the source code.