Fiddle with the test number for util/fdflag
This test case prints two ok lines behind the back of the TAP library
in a child process.  This means that the status report at the end of
the test was wrong.  Increment testnum so at least it's correct if
everything passes, although the status report will still miss failures.