Remove the NEED_KAFS conditional from m4/kafs.m4
Don't set the Automake conditional directly in m4/kafs.m4, since the
kafs probing may be conditional on whether AFS support is desired.
Leave it to the configure.ac file to set the Automake conditional.

Also remove the KAFS_SYSCALL_OBJECTS substitution variable, which is
no longer used now that we include the relevant files in kafs/kafs.c.