Add generic buffer handling utility functions
Add a new set of utility functions for managing a data buffer.  A
buffer dynamically resizes and has a concept of used and unused data,
making it suitable for buffering and processing data from an input
stream.  It has associated functions for adding data to the buffer in
various ways.  Taken from work Russ Allbery did for INN's library,
based on buffer code originally found in innd.