Initialize result variables before calling krb5_appdefault_string
Mac OS X Lion ships with a Kerberos library that stubs out the
krb5_appdefault_* functions to do nothing except warn in syslog.
This caused possible random failures in the configuration parsing,
since the code was assuming that krb5_appdefault_string would always
set the result variable (possibly to NULL).  Initialize it to avoid
this problem.