Stop casting PAM flags to unsigned int
This cast was required to avoid a -Wconversion warning on platforms
like Linux where PAM flags are defined as unsigned, but adding the
cast to only one side of the equality causes a more serious comparison
warning on platforms where they're signed.  We can't use -Wconversion
right now anyway, so instead of going down the path of adding casts to
every use of a PAM flag, just let the signed flag value be promoted to
unsigned.