Abort during configure if kafs/sys-syscall.c is used w/o AFS headers
Abort during configure if building the replacement kafs library on a
platform that uses regular system calls and afs/param.h is not found.
The system call interface can't be built without the AFS headers, and
it's better to fail with a descriptive message at configure time than
with an obscure message at compile time.

Remove a duplicate check for a header in m4/kafs.m4.

Include afs/param.h unconditionally in sys-syscall.c.  There's no
point in proceding if we don't have the header.

Change the license of sys-syscall.c to public domain, since it's too
embarassingly simple to put a real license on (and probably isn't
copyrightable anyway).