Cast struct sockaddrs through void * for -Wcast-align
Add an additional cast through void * when casting struct sockaddrs
to another, more specific sockaddr type.  This suppresses gcc warnings
from -Wcast-align and allows that warning flag to be used.  gcc
doesn't have a way of knowing that the original source of the sockaddr
would have allocated it with appropriate alignment.