Add opaque data to is_function_output, add new run_setup function
Improve the is_function_output TAP add-on interface to take an opaque
data pointer and pass it into the called function.

Add a new run_setup TAP add-on function that runs a given command and
calls bail if it fails, used for doing test setup in an external
command.  (Some things are easier to do in shell than in C.)