Fix error handling in connect with timeout, fix test suite
In the network utility library, when a connection with a timeout
concludes, properly store the recovered error code in errno where the
caller can pick it up.  Fix some problems with the test suite for
connect with timeout on systems where this cannot be tested by using a
short listen queue.