Improve Kerberos probe header portability
In the Kerberos library probes, probe for both krb5.h and krb5/krb5.h
and use the latter if the former isn't found.  Add an
RRA_INCLUDES_KRB5 macro that expands into the headers to use for
Kerberos probes.  This allows the header to be found on NetBSD
systems.  Thanks to Fredrik Pettai for the report.

Also use [] for empty arguments to configure macros and use more modern
macros for RRA_FUNC_KRB5_GET_INIT_CREDS_OPT_FREE_ARGS.