Add probe and portability for kadmin, password expiration utility
Add a probe for Kerberos kadmin client libraries.  This uses a new
framework for library probes (m4/lib-helper.m4 and m4/krb5-config.m4)
that may eventually be used for the other library probes.  Add a
portability wrapper around kadm5/admin.h and some example probes in
configure.ac.

Add a Kerberos TAP utility function that uses the kadmin client
library to set a password expiration time on a principal for software
that needs to test behavior of expired passwords.