Cleanup and simplification of Kerberos TAP add-on
Add reading of Kerberos username and password configuration to
kerberos_setup in the Kerberos TAP add-on and always return a struct
of configuration.  Return the path to the keytab configuration, if
present.  Remove the alternative implementation of the Kerberos
utility functions that doesn't require the Kerberos libraries and
instead use #ifdef inside the primary implementation to handle this.