Eliminate the bool arguments in the buffer_sprintf interface
Remove the bool arguments to buffer_sprintf and buffer_vsprintf and
instead introduce new buffer_append_sprintf and _vsprintf functions to
append to the buffer, which is what the functions did with a true
argument.  This avoids having a bool argument to functions, the
meaning of which is often obscure in the calling code.