Separate password and authtok PAM testing
In the PAM testing framework, don't copy the provided password into
the PAM_AUTHTOK item.  Instead, take new configuration parameters
authtok and oldauthtok that, if provided, are copied into the PAM data
before running the script.  This allows separate testing of a saved
password and a password provided at a prompt.  Document in the fakepam
README file all of the possible configuration parameters.