Update Perl test scripts to Term::ANSIColor versions
Change the license on all of the Perl test files to public domain to
make them easier to copy from one place to another.  Add the examples
directory to Test::Perl::Critic, Test::Pod, and Test::Pod::Spelling
tests.  Provide an easy way to add additional private symbols to the
Test::Pod::Coverage test.  Add coverage testing.  Add a test script
for Test::Synopsis.