Update network bind functions for socket type, return value
The network_bind_* functions now take a socket type as an additional
argument so that they can be used with UDP-based services.

network_bind_all now returns a boolean, which will be false if no
sockets could be bound due to some error.  Callers may check this
instead of checking if the socket count is zero.