Add new network_wait_any function
New network_wait_any function that waits for an array of file
descriptors and returns the first one that selects ready for read.
This can be used by UDP servers similar to network_accept_any for TCP
servers.

Add a test of UDP handling to the network server test, and also test
the new network_wait_any function.