Memory allocation and deallocation coding style cleanup
Don't check whether a pointer is NULL before passing it into free
and instead assume free can handle NULL pointers properly.  This
has been true for many years.

Fix buffer_free and the vector_free in pam-util to support taking
NULL pointers and doing nothing with them.

Allocate memory with calloc and assume this sets pointers to NULL
instead of explicitly initializing them.  We already had to assume
this in various places, and architectures where the all-zero bit
pattern is not the NULL pointer are exceedingly rare.