Skip Perl critic tests if directory is not writable
Perl::Tidy 20130922 unconditionally tries to create a log file in
the current directory if not told to put it somewhere else, the
current Perl::Critic does not tell it to put it somewhere else, and
this fails when the source directory is not writable (such as via
make distcheck).  In that case, skip the test.