Replace malloc using multiplication with calloc
Use calloc in preference to calculating a malloc size with
multiplication everywhere.  In most places this caution was probably
not necessary, but uniformity is easier to audit and no one will ever
notice the speed difference between malloc and calloc.

portable/system.h now also guarantees the inclusion of inttypes.h
(which it had included for years as an implementation detail) and
limits.h (to ensure the availability of SIZE_TYPE on older systems).

Also make xwritev more robust by doing range checking on iovcnt
early.  It's unlikely this would have caused a real problem anywhere,
but this is more obviously correct.