Add UNUSED annotation to two internal network functions
If a system doesn't support IPv6, but does support some of the
IPv6 constants (such as when IPv6 is disabled for some reason),
this would produce warnings about functions that are defined but
not used.  This should be quite rare, but add UNUSED annotations
to the functions anyway, since there should be no harm.