Conditionalize include of sys/time.h and add time.h for Windows
Windows doesn't have sys/time.h and defines struct timeval in
time.h.  Conditionalize the inclusion of sys/time.h for Windows
and add an include of time.h where it was missing.

This is mostly pointless, since most of these files won't build
meaningfully on Windows anyway, but may as well be consistent.