Consistently #undef functions before defining them for testing
When testing portability functions, we redefine the function to
a different name to not conflict with system headers.  If the
function is implemented by the system as a macro, we have to
undefine it first.  This was being done in some places but not in
others, causing compliation problems with inet_aton on FreeBSD.
Do this consistently everywhere.