In network server test, try connecting to IPv6 socket
Check that we can actually connect to the socket we created, since
I was having problems with Travis-CI supporting socket creation but
then not allowing us to connect to the socket.  Skip the IPv6 tests
if we can't connect to the resulting socket.

This runs the risk of skipping the test instead of failing if we
actually break either network_bind_ipv6 or network_connect_host, but
I can't think of a better way of handling this.