Add new macros for checking the properties of Perl
Add a new RRA_PROG_PERL macro that finds the path of a Perl
interpreter of the given version or newer.  Add a new
RRA_PERL_CHECK_MODULE macro that checks whether a given module can be
loaded by Perl.  Add a new RRA_LIB_PERL macro that finds the compiler
and linker flags for embedding Perl.  These are based on the configure
macros used in INN.