Fix return value checks for snprintf
If the buffer is insufficient size, snprintf returns the total
number of bytes that would be written, minus the trailing nul
byte.  This means there was an off-by-one error in several checks
of the return value, where a buffer exactly one byte too small
would result in a return status interpreted as success.

In all these places, this case appears to be impossible for other
reasons, but patch this up just in case.

(Patch originally submitted against INN.)