Increase the buffer size of the network write timeout test
The default Debian cloud images use a 12MB TCP sending buffer,
which meant that the timeout test never triggered because the
entire write of 8MB was absorbed by the buffer.  Bump the buffer
size to 15MB and add a comment explaining why that number was
chosen, with a reference to the Debian bug where the Ruby
maintainers sorted this out.

Many thanks to Lucas Nussbaum for investigating this and to
Eric Wong for sorting out why this buffer was causing tests to
fail.