Only test truth of socket flags
When testing network_set_reuseaddr and related functions, don't
require that the option be set to 1 explicitly, just that it has
a true value.  Debian/kFreeBSD returns the numeric value of the
socket option in some cases when the socket option is set rather
than returning 1, and POSIX just documents that this is a boolean
flag (so only the truthfulness of the value matters).