Properly handle signed overflow in run_child_function
The previous code attempted to handle signed overflow when
nul-terminating the output buffer of run_child_function, but didn't
handle it in the actual loop.  Prompted by the gcc warning with -O2,
do what we should have done originally and actually check against
SSIZE_MAX before increasing the count of read bytes, bailing out of
the test if we would overflow.