Suppress krb5-config probe output with --with-krb5
If --with-krb5 (or the equivalent for other probes) is given and a
krb5-config was found at the install root, skip the PATH search for
krb5-config to avoid confusing output.  This results in redoing the
probe and using cached output when testing multiple krb5-config
features, but that seems better than the confusing output.

Reported by Julien ÉLIE.