Check results of RRA_LIB_PERL and RRA_LIB_PYTHON
The Autoconf RRA_LIB_PERL macro now checks for the EXTERN.h header and
for perl_alloc in the libperl library, to double-check that the
discovered linkage information actually works.  (It may not on, for
example, a Debian host without libperl-dev installed.)

Do the same for RRA_LIB_PYTHON with Python.h and Py_Initialize.