Check headers as well as libraries in Autoconf probes
The Autoconf RRA_LIB_BDB, RRA_LIB_OPENSSL, RRA_LIB_SASL, and
RRA_LIB_ZLIB macros, and their _OPTIONAL counterparts, all now check
that the library's key header can be found as well as the library.
The _OPTIONAL versions treat the library as missing if the header is
not present; the non-optional versions will abort configure with an
error if the header is not present.