Fix RRA_LIB_KRB5_OPTIONAL with --enable-reduced-depends
Fix RRA_LIB_KRB5_OPTIONAL probing in the --enable-reduced-depends case
when linking with libkrb5 is not possible but libcom_err is present.
The probes for com_err were previously not skipped, resulting in a
non-empty KRB5_LIBS containing only -lcom_err, which incorrectly
caused the macro to indicate that Kerberos libraries were present.