Close some memory leaks in util/network/server test
The code allocated a sockaddr to get the address family but didn't
explicitly free it in the children of a fork, which valgrind is now
detecting as a memory leak.  Since we only care about the family,
copy it out of the struct and free it immediately.

Explicitly die in the child if the address family is unknown.