Improve CI installation of Perl modules
Use perl-actions/install-with-cpanm, which among other things
disables testing, which was part of the reason for slow install
times before.  Externalize the modules to install in a cpanfile
and move all the Perl packages other than those required to
support cpanm into the cpanfile and away from Debian packages.

This introduces a bit of a bias towards supporting current systems,
instead of testing with Perl on older Ubuntu, but I think that's
acceptable.

Add caching of the Perl modules.  This requires installing them
into the local::lib path instead of the system path and setting
up appropriate Perl environment variables in ci/test.

Move the Debian packages into a separate file that's parsed and
used by ci/install, to further move towards making this workflow
infrastructure generic.