Work around NetBSD prototype of reallocarray
NetBSD prefers reallocarr to reallocarray and only prototypes the
latter if _OPENBSD_SOURCE is defined, which is not one of the macros
Autoconf defines by default.  Work around this for now by checking
whether reallocarray is declared, not only whether it's available, and
protyping it if it's not declared.  This uses the semi-hidden
deprecated symbol on NetBSD.