Update perlcritic configuration
Update perlcritic configuration to remove (instead of tweaking) two
warnings that I keep running into and have never found useful, and to
cope with the renaming of Perl::Critic::Freenode to
Perl::Critic::Community.  The latter requires a workaround in the test
programs because of perlcritic's unsuppressible warning about unknown
policies.