Request posix_prefix paths from Python sysconfig
When using the Python sysconfig module to determine paths,
explicitly request the posix_prefix scheme.  This is the default
in current versions of Python, but in Python 2.7 as packaged in
Debian buster, the default scheme is posix_local.  This breaks
determination of the include path for Python.h, since it instead
returns the include path used for locally-installed modules.

Explicitly requesting the posix_prefix scheme avoids this problem.