Rework tests/util/fakewrite.c code
cppcheck got confused by this code. Those reasons were probably a
bug, but the code was using magical constants and was a bit awkward.
Rework it to use sizeof and make the analysis simpler so that
cppcheck won't get confused. Also return ENOSPC for writes past the
end of the buffer, just for cleanliness even though the test suite
doesn't do that.