Do not unlink temporary file in the mkstemp test
Do not unlink the file and then write to it in the mkstemp
replacement test suite, since this fails on some file systems. This
is not ideal since the calling program may be relying on traditional
UNIX file semantics, but it can be addressed by moving temporary
files to another file system and doesn't indicate a bug in the
replacement function. Thanks to Julien ÉLIE for the report.